added maximum_independent_set in the clique module and and updated tests
