+++ networkx/algorithms/approximation/__init__.py
@@ -14,7 +14,6 @@ from networkx.algorithms.approximation.clique import *
 from networkx.algorithms.approximation.connectivity import *
 from networkx.algorithms.approximation.dominating_set import *
 from networkx.algorithms.approximation.kcomponents import *
-from networkx.algorithms.approximation.independent_set import *
 from networkx.algorithms.approximation.matching import *
 from networkx.algorithms.approximation.ramsey import *
 from networkx.algorithms.approximation.steinertree import *
